Blinking food
Added gold border and blinking animation to the food, making it easier for the user to find the food when it appears, and it appears like it's "spawning" in.

If you like, and it follows the rules, please accept for hacktoberfest2020. I'm not trying to spam, so if it seems like it, let me know, I'd be happy to make more changes to make this eligible. :)